requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudevents==1.11.0
colorama==0.4.6
cryptography==43.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
fastapi==0.115.4
ffmpy==0.4.0
filelock==3.16.1
flexcache==0.3
flexparser==0.3.1
frozenlist==1.5.0
fsspec==2024.10.0
gradio==5.4.0
gradio_client==1.4.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.4.0
isodate==0.7.2
Jinja2==3.1.4
jiter==0.7.0
jmespath==1.0.1
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langsmith==0.1.139
lazy-object-proxy==1.10.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.23.1
mdurl==0.1.2
more-itertools==10.5.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.53.0
openapi-core==0.19.4
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
orjson==3.10.11
packaging==24.1
pandas==2.2.3
parse==1.20.2
pathable==0.4.3
pillow==11.0.0
Pint==0.24.3
portalocker==2.10.1
prance==23.6.21.0
propcache==0.2.0
pybars4==0.9.13
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
PyMeta3==0.5.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rpds-py==0.20.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
ruff==0.7.2
safehttpx==0.1.1
semantic-kernel==1.13.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.2
tenacity==9.0.0
tiktoken==0.8.0
tomlkit==0.12.0
tqdm==4.66.6
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==12.0
Werkzeug==3.1.2
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
